+++ src/main/java/com/nokia/seatmanagement/core/db/repository/SeatRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.nokia.seatmanagement.models.Seats;  
 
-public interface SeatRepository extends CrudRepository<Seats, Integer> {
+public interface SeatRepository extends CrudRepository<Seats, Long> {
 
 }
